Create super class Contact, and inherited by Address, Phone, and Email classes.
